redesign light & dark themes, inspired by Qt's Neptune3 UI
