+++ application/resources/resource.qrc
@@ -10,9 +10,13 @@
     <qresource prefix="/icons">
         <file>folder.png</file>
         <file>arrow_right.png</file>
+        <file>arrow_right_dark.png</file>
         <file>arrow_down.png</file>
+        <file>arrow_down_dark.png</file>
         <file>timer_start.png</file>
+        <file>timer_start_dark.png</file>
         <file>timer_stop.png</file>
+        <file>timer_stop_dark.png</file>
         <file>redo.png</file>
         <file>undo.png</file>
         <file>clock.png</file>
@@ -25,5 +29,7 @@
         <file>appicon.png</file>
         <file>appicon_32.png</file>
         <file>appicon_64.png</file>
+        <file>handle.png</file>
+        <file>handle_urgent.png</file>
     </qresource>
 </RCC>
