+++ application/resources/stylesheet.css
@@ -228,6 +228,7 @@ OverlayWidget #report {
 
 TaskListWidget {
   background: transparent;
+  font-size: 8pt; 
 }
 
 TaskWidget, TaskWidget #pShadow {
@@ -242,10 +243,11 @@ TaskWidget #pFrame{
 
 TaskWidget #pTitle {
   qproperty-drawOutline: true;
-  padding-left: 4px;
-  font-size: 11pt;
+  qproperty-drawShadow: false;
+  padding-left: 4px; 
   color: rgba(255, 255, 255, 255);
-}
+  font-size: 8pt;  
+} 
 
 TaskWidget #pLinks {
   border: 0px solid #ff9900;
@@ -374,11 +376,7 @@ LinkWidget:focus {
   padding: 0 0 6px 0;
   border: 0px solid red;
 }
-
-#pDynamicProperties QLabel {
-  font-size: 11pt;
-}
-
+ 
 #pSeparator {
   border-top: 1px solid qlineargradient(x1:0, y1:0, x2:1, y2:0
                 stop:0 rgba(0, 0, 0, 0),
